hzCmd-bench

+++ src/main/java/hz/map/GetAsyncWait.java
@@ -9,12 +9,11 @@ import java.util.concurrent.TimeoutException;
 
 public class GetAsyncWait extends MapBench {
 
-    public long timeout=0;
+    public int timeout=0;
 
     public void timeStep() throws InterruptedException, ExecutionException, TimeoutException {
         int k = random.nextInt(keyDomain);
         ICompletableFuture f = map.getAsync(k);
-
         f.get(timeout, TimeUnit.MILLISECONDS);
     }
 }
